requirements.txt
numpy
pandas
pandas_ta
pytorch_lightning
torchmetrics
